Update delete endpoints priviliages

+++ src/main/java/edu/vutfit/ThirstQuest/controller/PhotoController.java
@@ -3,6 +3,7 @@ package edu.vutfit.ThirstQuest.controller;
 import edu.vutfit.ThirstQuest.model.Photo;
 import edu.vutfit.ThirstQuest.service.PhotoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
@@ -25,8 +26,21 @@ public class PhotoController {
     }
 
     @DeleteMapping("/{id}")
-    public void deletePhoto(@PathVariable UUID id) {
-        photoService.deletePhoto(id);
+    public String deletePhoto(@PathVariable UUID id, Authentication authentication) {
+        Photo photo = photoService.getPhotoById(id);
+        String currentUserEmail = authentication.getName();
+
+        if (photo == null) {
+            return "Review not found";
+        }
+
+        // Only allow the owner of the photo or an ADMIN to delete it
+        if (photo.getUser().getEmail().equals(currentUserEmail) || authentication.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+            photoService.deletePhoto(id);
+            return "Photo deleted";
+        }
+
+        return "Unauthorized to delete this photo";
     }
 }
